requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pdfplumber>=0.10.0
pikepdf>=8.0.0
pandas>=2.2.0
python-dateutil>=2.9.0
playwright>=1.42.0
aiosqlite>=0.20.0
pydantic>=2.6.0
sse-starlette>=2.0.0
python-multipart>=0.0.9
reportlab>=4.0.0
